# api/openapi-campuslens.yaml
openapi: 3.0.3
info:
  title: CampusLens API
  version: 1.0.0
  description: CampusLens（校园慧眼）校园地标智能检索与导览系统接口草案。
servers:
  - url: http://localhost:8080
    description: Spring Boot backend
paths:
...
  /api/guests:
    post:
      summary: 幂等分配浏览器游客身份
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestIdentityRequest"
      responses:
        "200":
          description: 返回数据库分配并由当前浏览器复用的游客编号
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestIdentityResponse"
        "400":
          description: clientToken 缺失或长度不合法
  /api/search/upload:
    post:
      summary: 提交异步图片检索任务
      tags: [Search]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
          description: 在同一登录用户或游客作用域内，同键同文件返回原任务；同键不同文件返回 409
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                guestId:
                  type: string
                  description: 未登录游客标识；登录用户身份由 Bearer token 解析
      responses:
        "202":
          description: 任务已进入 Redis 队列
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchJobSubmission"
        "400":
          description: 图片格式或参数错误
        "409":
          description: 同一幂等键用于不同图片
        "429":
          description: 队列容量已满，并返回 Retry-After
        "503":
          description: Redis 队列不可用，且不会保留伪成功记录
  /api/search/jobs/{jobId}:
    get:
      summary: 查询异步图片检索任务状态
      tags: [Search]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Search-Job-Token
          in: header
          required: false
          schema:
            type: string
          description: 游客任务必须携带；登录任务使用 Bearer token 校验所有权
      responses:
        "200":
          description: 当前任务状态及完成后的 Top-5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchJobStatus"
        "401":
          description: 登录用户无权访问，或游客任务令牌无效
        "404":
          description: 任务不存在
...
  /api/landmarks/{id}:
    get:
      summary: 获取地标详情
      tags: [Landmark]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 地标详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LandmarkDetail"
        "404":
          description: 地标不存在
  /api/feedback:
    post:
      summary: 提交识别反馈
      tags: [Feedback]
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "200":
          description: 反馈提交成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
  /api/auth/register:
    post:
      summary: 普通用户注册
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterRequest"
      responses:
        "200":
          description: 注册成功并返回用户身份
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: 用户名重复、密码过短或参数错误
  /api/auth/login:
    post:
      summary: 用户登录；admin/admin 会返回管理员身份
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: 登录成功；role=admin 时前端自动进入后台
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: 用户名或密码错误
  /api/me/account:
    get:
      summary: 查询当前登录账号资料
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 当前账号资料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountProfile"
        "401":
          description: 未登录、账号不存在或已停用
  /api/me/account/email:
    put:
      summary: 更新当前登录账号邮箱
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountEmailUpdateRequest"
      responses:
        "200":
          description: 邮箱更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountUpdateResponse"
        "400":
          description: 邮箱格式或长度不符合要求
        "401":
          description: 未登录
  /api/me/account/password:
    put:
      summary: 验证当前密码后修改登录密码
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountPasswordUpdateRequest"
      responses:
        "200":
          description: 密码修改成功，当前会话继续有效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountUpdateResponse"
        "400":
          description: 当前密码错误、新密码过短或与当前密码相同
        "401":
          description: 未登录
  /api/me/account/avatar:
    post:
      summary: 上传当前登录账号的裁剪头像
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: JPG 或 PNG，最大 5 MB，尺寸 32×32 至 4096×4096
      responses:
        "200":
          description: 头像保存成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountUpdateResponse"
        "400":
          description: 文件为空、格式伪造、尺寸或大小不符合要求
        "401":
          description: 未登录
  /api/me/search-records:
    get:
      summary: 查询当前登录用户个人检索历史
      tags: [User]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        "200":
          description: 当前登录用户自己的检索记录
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSearchRecord"
        "401":
          description: 未登录
  /api/check-ins:
    get:
      summary: 查询校园打卡留言
      tags: [CheckIn]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: landmarkId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: guestId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: 打卡留言列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckInRecord"
    post:
      summary: 发布校园打卡留言
      tags: [CheckIn]
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInRequest"
      responses:
        "200":
          description: 发布成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInRecord"
        "400":
          description: 检索记录状态不可用或地标不属于该记录 Top-5
        "401":
          description: 检索记录不属于当前登录用户或游客
        "409":
          description: 该检索记录已经发布过打卡
  /api/check-ins/{id}:
    get:
      summary: 查询单条校园动态及完整回复树
      tags: [CheckIn]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: guestId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: 帖子详情；replies 为完整嵌套回复树
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInRecord"
        "400":
          description: 帖子不存在或已隐藏
  /api/check-ins/{id}/like:
    post:
      summary: 点赞或取消点赞打卡留言
      tags: [CheckIn]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: guestId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: 点赞状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /api/check-ins/{id}/replies:
    post:
      summary: 回复打卡留言
      tags: [CheckIn]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInReplyRequest"
      responses:
        "200":
          description: 回复成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInReply"
        "400":
          description: 帖子不存在，或 parentReplyId 不属于当前帖子
  /api/check-in-replies/{id}/like:
    post:
      summary: 点赞或取消点赞回复
      tags: [CheckIn]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: guestId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: 回复点赞状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyLikeResponse"
  /api/admin/auth/login:
    post:
      summary: 管理员轻量登录（兼容旧接口，前端不直接展示）
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: 登录成功；V2 前端使用 /api/auth/login 隐式进入后台
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "400":
          description: 账号或密码错误
  /api/admin/search-records:
    get:
      summary: 查看最近检索记录
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 最近 50 条检索记录
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminSearchRecord"
  /api/admin/algorithm/runtime:
    get:
      summary: 查看算法运行状态
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 主算法实例、模型、SAR、活动索引和重建任务状态
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/admin/landmarks:
    post:
      summary: 新增地标
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LandmarkCreateRequest"
      responses:
        "200":
          description: 新增成功
  /api/admin/landmarks/{id}:
    put:
      summary: 修改地标
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LandmarkCreateRequest"
      responses:
        "200":
          description: 修改成功
  /api/admin/landmarks/{id}/images:
    post:
      summary: 上传地标样本图片
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                angle:
                  type: string
                lightCondition:
                  type: string
      responses:
        "200":
          description: 上传成功
  /api/admin/index/rebuild:
    post:
      summary: 重建地标统计参数
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "202":
          description: 已提交索引重建任务
          content:
            application/json:
              schema:
                type: object
                required: [rebuildJobId, status]
                properties:
                  rebuildJobId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [building]
                  createdAt:
                    type: string
                    format: date-time
                  error:
                    type: string
                    nullable: true
  /api/admin/index/rebuild/{jobId}:
    get:
      summary: 查询索引重建任务状态
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 重建任务状态；完成时包含 indexVersion，失败时包含 error
          content:
            application/json:
              schema:
                type: object
                required: [rebuildJobId, status]
                properties:
                  rebuildJobId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [building, switching, completed, failed]
                  indexVersion:
                    type: string
                    nullable: true
                  error:
                    type: string
                    nullable: true
  /api/admin/feedback:
    get:
      summary: 查看反馈记录
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 反馈列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminFeedbackRecord"
  /api/admin/feedback/{id}:
    get:
      summary: 查看反馈详情
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 反馈详情，含上传图、Top-5 快照和校正样本状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFeedbackDetail"
  /api/admin/feedback/{id}/status:
    post:
      summary: 更新反馈处理状态
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackStatusRequest"
      responses:
        "200":
          description: 状态更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "400":
          description: 任务尚未成功完成，或候选地标不属于本次 Top-5
        "401":
          description: 登录用户或游客标识与检索任务所有者不一致
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 登录或注册接口返回的服务端会话 token。后台接口要求 role=admin。
  schemas:
    GuestIdentityRequest:
      type: object
      required: [clientToken]
      properties:
        clientToken:
          type: string
          minLength: 16
          maxLength: 200
    GuestIdentityResponse:
      type: object
      required: [guestId]
      properties:
        guestId:
          type: string
          pattern: '^guest#[1-9][0-9]*$'
    SearchJobSubmission:
      type: object
      required: [jobId, searchRecordId, status, jobToken, pollAfterMs]
      properties:
        jobId:
          type: string
          format: uuid
        searchRecordId:
          type: integer
          format: int64
        status:
          type: string
          enum: [queued, processing, success, low_confidence, failed]
        jobToken:
          type: string
          description: 游客轮询令牌；数据库仅保存 SHA-256 哈希
        pollAfterMs:
          type: integer
          example: 1000
    SearchJobStatus:
      type: object
      required: [jobId, searchRecordId, status, attemptCount, lowConfidence, results]
      properties:
        jobId:
          type: string
          format: uuid
        searchRecordId:
          type: integer
          format: int64
        status:
          type: string
          enum: [queued, processing, success, low_confidence, failed]
        attemptCount:
          type: integer
        uploadImageUrl:
          type: string
        guestId:
          type: string
        lowConfidence:
          type: boolean
        message:
          type: string
        errorCode:
          type: string
          nullable: true
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        queuedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
...
    LandmarkDetail:
      allOf:
        - $ref: "#/components/schemas/LandmarkSummary"
        - type: object
          properties:
            description:
              type: string
            locationText:
              type: string
            mapX:
              type: number
            mapY:
              type: number
            images:
              type: array
              items:
                $ref: "#/components/schemas/LandmarkImage"
...
    FeedbackRequest:
      type: object
      required: [searchRecordId, feedbackType]
      properties:
        searchRecordId:
          type: integer
          format: int64
        predictedLandmarkId:
          type: integer
          format: int64
        confirmedLandmarkId:
          type: integer
          format: int64
        guestId:
          type: string
          nullable: true
          description: 未登录游客标识，必须与检索任务一致；登录用户身份由 Bearer token 解析
        feedbackType:
          type: string
          enum: [correct, wrong, uncertain]
        comment:
          type: string
...
    AuthLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: student01
        password:
          type: string
          example: password123
...